refactor(cli): split argumentsChecker.py into src/cli/{parser,validator,startup}
2256-line file split by concern:
- src/cli/parser.py    -- DidYouMeanArgumentParser, TASHelpFormatter,
                          _buildParser, createParser, all _add*Options,
                          capabilityMethods, _listMethods, str2bool
- src/cli/validator.py -- argumentsChecker, _configureProcessingSettings,
                          _adjustMethodsBasedOnCuda, _validateCustomUpscaleModel,
                          processURL, and the other validation helpers
- src/cli/startup.py   -- _handleDependencies, _promptDownloadRequirementsSelection

src/utils/argumentsChecker.py becomes a re-export shim so all existing
import sites (main.py, tests, dependencyHandler) continue to work unchanged.
44 argumentsChecker tests pass.
