+++ src/utils/argumentsChecker.py
@@ -1,2256 +1,81 @@
 """
-Command Line Argument Parser
+Backward-compatibility shim. Import from src.cli.* directly.
 
-Handles parsing and validation of command line arguments for The Anime Scripter.
-Provides comprehensive argument definitions and validation logic.
+  src.cli.parser    -- argparse classes, _buildParser, createParser, _add*Options
+  src.cli.validator -- argumentsChecker, validation helpers
+  src.cli.startup   -- _handleDependencies, _promptDownloadRequirementsSelection
 """
 
-import os
-import logging
-import sys
-import argparse
-import shutil
-import src.constants as cs
-
-from src.version import __version__
-
-_logAndPrint = None
-
-
-def logAndPrint(message, colorFunc="cyan", level="INFO"):
-    global _logAndPrint
-    if _logAndPrint is None:
-        from src.utils.logAndPrint import logAndPrint as _lap
-        _logAndPrint = _lap
-    _logAndPrint(message, colorFunc, level)
-
-
-class DidYouMeanArgumentParser(argparse.ArgumentParser):
-    """
-    Custom ArgumentParser that provides "did you mean?" suggestions for invalid choices.
-    
-    When a user provides an invalid choice for an argument, this parser will
-    suggest similar valid choices based on string similarity.
-    """
-
-    def _levenshteinDistance(self, s1, s2):
-        """
-        Calculate the Levenshtein distance between two strings.
-        
-        Args:
-            s1: First string
-            s2: Second string
-            
-        Returns:
-            int: The edit distance between the two strings
-        """
-        if len(s1) < len(s2):
-            return self._levenshteinDistance(s2, s1)
-        
-        if len(s2) == 0:
-            return len(s1)
-        
-        previousRow = range(len(s2) + 1)
-        for i, c1 in enumerate(s1):
-            currentRow = [i + 1]
-            for j, c2 in enumerate(s2):
-                insertions = previousRow[j + 1] + 1
-                deletions = currentRow[j] + 1
-                substitutions = previousRow[j] + (c1 != c2)
-                currentRow.append(min(insertions, deletions, substitutions))
-            previousRow = currentRow
-        
-        return previousRow[-1]
-
-    def similarityScore(self, invalidValue, choice):
-        """
-        Calculate a similarity score between an invalid value and a valid choice.
-        Higher score = better match.
-        
-        Args:
-            invalidValue: The invalid value provided by the user
-            choice: A valid choice to compare against
-            
-        Returns:
-            float: Similarity score (higher is better)
-        """
-        invalidLower = invalidValue.lower()
-        choiceLower = choice.lower()
-        
-        distance = self._levenshteinDistance(invalidLower, choiceLower)
-        maxLen = max(len(invalidLower), len(choiceLower))
-        baseSimilarity = 1.0 - (distance / maxLen) if maxLen > 0 else 0
-        
-        bonus = 0.0
-        
-        if choiceLower.startswith(invalidLower) or invalidLower.startswith(choiceLower):
-            bonus += 0.4
-        
-        commonPrefixLen = 0
-        for i in range(min(len(invalidLower), len(choiceLower))):
-            if invalidLower[i] == choiceLower[i]:
-                commonPrefixLen += 1
-            else:
-                break
-        bonus += (commonPrefixLen / max(len(invalidLower), len(choiceLower))) * 0.2
-        
-        if invalidLower in choiceLower:
-            bonus += 0.15
-
-        # Family awareness: reward sharing the base name before a backend
-        # suffix (e.g. 'span-trt' -> 'span-tensorrt' over 'span-ncnn').
-        if "-" in invalidLower and "-" in choiceLower:
-            if invalidLower.rsplit("-", 1)[0] == choiceLower.rsplit("-", 1)[0]:
-                bonus += 0.15
-        
-        return baseSimilarity + bonus
-
-    def getSuggestions(self, invalidValue, validChoices, maxSuggestions=5):
-        """
-        Get suggestions for an invalid value from a list of valid choices.
-        
-        Args:
-            invalidValue: The invalid value provided by the user
-            validChoices: List of valid choices
-            maxSuggestions: Maximum number of suggestions to return
-            
-        Returns:
-            List of suggested valid choices, sorted by similarity
-        """
-        # Score all choices
-        scoredChoices = [
-            (choice, self.similarityScore(invalidValue, choice))
-            for choice in validChoices
-        ]
-        
-        # Sort by score (descending) and then by length (ascending) for ties
-        scoredChoices.sort(key=lambda x: (-x[1], len(x[0])))
-        
-        # Return top suggestions (only if they have a reasonable score)
-        suggestions = []
-        for choice, score in scoredChoices:
-            if score > 0.3 and len(suggestions) < maxSuggestions:  # Minimum threshold
-                suggestions.append(choice)
-        
-        return suggestions
-
-    def _palette(self):
-        """ANSI color codes, blanked out when the terminal lacks color."""
-        useColor = _supportsColorStdout()
-        code = lambda c: c if useColor else ""
-        return {
-            "RED": code("\x1b[1;31m"),
-            "YELLOW": code("\x1b[1;33m"),
-            "GREEN": code("\x1b[1;32m"),
-            "CYAN": code("\x1b[1;36m"),
-            "DIM": code("\x1b[2m"),
-            "RESET": code("\x1b[0m"),
-        }
-
-    def _collectOptionStrings(self):
-        """Every registered option string (e.g. '--upscale', '-v')."""
-        options = []
-        for action in self._actions:
-            options.extend(action.option_strings)
-        return options
-
-    def _suggestOptions(self, invalidOption, validOptions, maxSuggestions=3):
-        """
-        Suggest the closest valid option names for a misspelled flag.
-
-        Compares only within the same dash style ('--' vs '-') and scores on
-        the dash-stripped core, so the shared leading dashes don't inflate the
-        similarity of every candidate.
-        """
-        isLong = invalidOption.startswith("--")
-        invalidCore = invalidOption.lstrip("-")
-
-        scored = []
-        for option in validOptions:
-            if option.startswith("--") != isLong:
-                continue
-            scored.append((option, self.similarityScore(invalidCore, option.lstrip("-"))))
-
-        scored.sort(key=lambda x: (-x[1], len(x[0])))
-        return [option for option, score in scored if score > 0.4][:maxSuggestions]
-
-    def _suggestOptionNames(self, message):
-        """
-        Provide "did you mean?" suggestions for unrecognized option names.
-
-        Returns True when it printed a tailored message (and exited); returns
-        False to let the default argparse handler take over (e.g. when the
-        unrecognized tokens are stray positionals, not flags).
-        """
-        tokens = message[len("unrecognized arguments:"):].split()
-        badOptions = [t for t in tokens if t.startswith("-") and len(t) > 1]
-        if not badOptions:
-            return False
-
-        validOptions = self._collectOptionStrings()
-        c = self._palette()
-
-        print(
-            f"\n  {c['RED']}Error:{c['RESET']} unrecognized "
-            f"{'option' if len(badOptions) == 1 else 'options'}: "
-            f"{c['RED']}{' '.join(badOptions)}{c['RESET']}",
-            file=sys.stderr,
-        )
-
-        for bad in badOptions:
-            core = bad.split("=", 1)[0]
-            suggestions = self._suggestOptions(core, validOptions)
-            if suggestions:
-                print(
-                    f"\n  {c['YELLOW']}Did you mean:{c['RESET']} "
-                    f"{c['GREEN']}{', '.join(suggestions)}{c['RESET']} "
-                    f"{c['DIM']}(instead of '{core}'){c['RESET']}",
-                    file=sys.stderr,
-                )
-
-        print(
-            f"\n  {c['CYAN']}Tip:{c['RESET']} {c['DIM']}run "
-            f"'main.py --help' for the full option list.{c['RESET']}",
-            file=sys.stderr,
-        )
-
-        sys.exit(2)
-
-    def error(self, message):
-        """
-        Override error method to provide "did you mean?" suggestions.
-        
-        Args:
-            message: The original error message from argparse
-        """
-        # Check if this is an invalid choice error
-        if "invalid choice:" in message and "choose from" in message:
-            import re
-
-            match = re.search(
-                r"argument (--?[\w-]+):\s*invalid choice:\s*'([^']+)'\s*\(choose from\s*(.+)\)",
-                message
-            )
-
-            if match:
-                argName = match.group(1)
-                invalidValue = match.group(2)
-                choicesStr = match.group(3)
-
-                # argparse wraps each choice in quotes ('a', 'b'); strip them
-                # so repr() below doesn't double-quote (e.g. "'span-tensorrt'").
-                choices = [c.strip().strip("'\"") for c in choicesStr.split(',') if c.strip()]
-
-                if choices:
-                    suggestions = self.getSuggestions(invalidValue, choices)
-
-                    useColor = _supportsColorStdout()
-                    RED = "\x1b[1;31m" if useColor else ""
-                    YELLOW = "\x1b[1;33m" if useColor else ""
-                    GREEN = "\x1b[1;32m" if useColor else ""
-                    CYAN = "\x1b[1;36m" if useColor else ""
-                    DIM = "\x1b[2m" if useColor else ""
-                    RESET = "\x1b[0m" if useColor else ""
-
-                    print(
-                        f"\n  {RED}Error:{RESET} argument {argName}: "
-                        f"invalid choice: {RED}'{invalidValue}'{RESET}",
-                        file=sys.stderr,
-                    )
-
-                    if suggestions:
-                        print(
-                            f"\n  {YELLOW}Did you mean:{RESET} "
-                            f"{GREEN}{', '.join(repr(s) for s in suggestions)}{RESET}",
-                            file=sys.stderr,
-                        )
-
-                    displayedChoices = choices[:10]
-                    choicesStrDisplay = ', '.join(repr(c) for c in displayedChoices)
-                    if len(choices) > 10:
-                        choicesStrDisplay += f", ... ({len(choices) - 10} more)"
-                    print(
-                        f"\n  {CYAN}Valid choices:{RESET} {DIM}{choicesStrDisplay}{RESET}",
-                        file=sys.stderr,
-                    )
-
-                    sys.exit(2)
-                    return
-        
-        # Misspelled option names (e.g. '--upscalee') -> suggest closest flag.
-        if message.startswith("unrecognized arguments:"):
-            self._suggestOptionNames(message)
-
-        # Everything else (missing/required args, bad values, ambiguous options):
-        # same minimal style as the suggestion paths -- no usage block, "Error:"
-        # prefix, leading blank line, indented to match.
-        c = self._palette()
-        print(f"\n  {c['RED']}Error:{c['RESET']} {message}", file=sys.stderr)
-        sys.exit(2)
-
-
-def _supportsColorStdout():
-    if os.environ.get("NO_COLOR"):
-        return False
-    if os.environ.get("FORCE_COLOR"):
-        return True
-    if not hasattr(sys.stdout, "isatty") or not sys.stdout.isatty():
-        return False
-    if sys.platform == "win32":
-        try:
-            return sys.getwindowsversion().build >= 14393
-        except Exception:
-            return False
-    return True
-
-
-class TASHelpFormatter(argparse.HelpFormatter):
-    """Compact, colored help formatter that groups long choice lists by backend."""
-
-    _INLINE_THRESHOLD = 6
-    _KNOWN_BACKENDS = ("tensorrt", "directml", "ncnn", "openvino")
-
-    def __init__(self, prog, indent_increment=2, max_help_position=36, width=None):
-        if width is None:
-            try:
-                width = os.get_terminal_size().columns
-            except (ValueError, OSError):
-                width = 100
-        super().__init__(
-            prog, indent_increment, max_help_position, min(max(width, 60), 120)
-        )
-
-    def _metavar_formatter(self, action, default_metavar):
-        if action.choices is not None:
-            choices = [str(c) for c in action.choices]
-            if len(choices) > self._INLINE_THRESHOLD:
-                result = action.dest.upper().rsplit("_", 1)[-1]
-            else:
-                result = "{%s}" % ", ".join(choices)
-
-            def format(tuple_size):
-                return (result,) * tuple_size
-
-            return format
-        return super()._metavar_formatter(action, default_metavar)
-
-    def _get_help_string(self, action):
-        help_text = action.help or ""
-        extras = []
-        if (
-            action.default is not None
-            and action.default is not argparse.SUPPRESS
-            and action.default not in (None, False, [], "")
-            and not isinstance(action.default, bool)
-            and action.option_strings
-            and "default:" not in help_text.lower()
-            and "default is" not in help_text.lower()
-        ):
-            if isinstance(action.default, list):
-                default_str = ", ".join(str(d) for d in action.default)
-            else:
-                default_str = str(action.default)
-            extras.append("default: %s" % default_str)
-        if extras:
-            help_text += " (%s)" % ", ".join(extras)
-        return help_text
-
-    def _format_action(self, action):
-        result = super()._format_action(action)
-        if not action.choices or len(list(action.choices)) <= self._INLINE_THRESHOLD:
-            return result
-        indent = " " * self._max_help_position
-        groups = self._group_choices([str(c) for c in action.choices])
-        lines = []
-        if len(groups) == 1:
-            # Single group - no backend labels needed
-            items_str = ", ".join(list(groups.values())[0])
-            lines.append(indent + items_str)
-        else:
-            for backend, items in groups.items():
-                label = "%-10s " % ("[%s]" % backend)
-                items_str = ", ".join(items)
-                lines.append(indent + label + items_str)
-        return result + "\n".join(lines) + "\n"
-
-    @classmethod
-    def _group_choices(cls, choices):
-        from collections import OrderedDict
-
-        groups = OrderedDict()
-        for choice in choices:
-            parts = choice.rsplit("-", 1)
-            if len(parts) == 2 and parts[1] in cls._KNOWN_BACKENDS:
-                backend = parts[1]
-            else:
-                backend = "cuda"
-            groups.setdefault(backend, []).append(choice)
-        return groups
-
-    def format_help(self):
-        text = super().format_help()
-
-        # Don't add header/colors for short outputs like --version
-        if "usage:" not in text.lower():
-            return text
-
-        # argparse reuses this formatter for usage-only output (format_usage,
-        # which print_usage calls on every error). The banner belongs only on
-        # the full --help screen, so skip it when there's just a usage block.
-        isUsageOnly = "\n\n" not in text.strip("\n")
-
-        color = _supportsColorStdout()
-
-        if isUsageOnly:
-            header = ""
-        elif color:
-            R = "\033[0m"
-            B = "\033[1m"
-            C = "\033[96m"
-            D = "\033[2m"
-            header = (
-                f"\n  {B}{C}The Anime Scripter{R} {D}v{__version__}{R}\n"
-                f"  {D}AI-powered video enhancement toolkit{R}\n\n"
-            )
-        else:
-            header = (
-                f"\n  The Anime Scripter v{__version__}\n"
-                f"  AI-powered video enhancement toolkit\n\n"
-            )
-
-        if not color:
-            return header + text
-
-        import re
-
-        R = "\033[0m"
-        B = "\033[1m"
-        C = "\033[96m"
-        G = "\033[92m"
-        Y = "\033[93m"
-
-        lines = text.split("\n")
-        out = []
-        for line in lines:
-            # Section headings like "General:" or "Video Processing:"
-            m = re.match(r"^(\s{0,2})([\w][\w /()-]*):(\s*)$", line)
-            if m:
-                out.append(f"{m.group(1)}{B}{C}{m.group(2)}{R}")
-                continue
-
-            # "Usage:" or "usage:" prefix
-            line = re.sub(r"^(usage:\s*|Usage:\s*)", f"{B}\\1{R}", line)
-
-            # --long-flags
-            line = re.sub(r"(--[\w][\w-]*)", f"{G}\\1{R}", line)
-            # -x short flags
-            line = re.sub(r"(?<=[\s,])(-[a-zA-Z])(?=[\s,\]])", f"{G}\\1{R}", line)
-
-            # {choices} in braces
-            line = re.sub(r"(\{[^}]+\})", f"{Y}\\1{R}", line)
-
-            out.append(line)
-
-        return header + "\n".join(out)
-
-
-def isAnyOtherProcessingMethodEnabled(args):
-    """
-    Check if any video processing operations are enabled.
-
-    Args:
-        args: Parsed command line arguments
-
-    Returns:
-        bool: True if any processing method is enabled
-    """
-    return any(
-        [
-            args.interpolate,
-            args.upscale,
-            args.segment,
-            args.restore,
-            args.stabilize,
-            args.resize,
-            args.dedup,
-            args.depth,
-            args.autoclip,
-            args.obj_detect,
-            args.moblur,
-        ]
-    )
-
-
-def str2bool(arg):
-    """
-    Convert string argument to boolean value.
-
-    Args:
-        arg: String or boolean argument
-
-    Returns:
-        bool: Converted boolean value
-
-    Raises:
-        argparse.ArgumentTypeError: If argument cannot be converted to boolean
-    """
-    if isinstance(arg, bool):
-        return arg
-    if arg.lower() in ("yes", "true", "t", "y", "1"):
-        return True
-    elif arg.lower() in ("no", "false", "f", "n", "0"):
-        return False
-    else:
-        raise argparse.ArgumentTypeError("Boolean value expected.")
-
-
-def _promptDownloadRequirementsSelection() -> str:
-    from inquirer import List, prompt
-
-    if cs.SYSTEM == "Darwin":
-        return "macos-mps"
-
-    currentPlatform = "windows" if cs.SYSTEM == "Windows" else "linux"
-    choices = [
-        (
-            "Full CUDA / TensorRT dependencies (GTX 16xx, RTX 20xx+, newer NVIDIA)",
-            f"{currentPlatform}-cuda",
-        ),
-        (
-            "Lite dependencies (GTX 10xx, AMD, Intel)",
-            f"{currentPlatform}-lite",
-        ),
-    ]
-    answers = prompt(
-        [
-            List(
-                "dependency_profile",
-                message=(
-                    f"Select which {currentPlatform.title()} dependencies to install for your hardware"
-                ),
-                choices=choices,
-            )
-        ]
-    )
-
-    if not answers:
-        logAndPrint("No dependency profile selected, exiting.", "red")
-        sys.exit()
-
-    return answers["dependency_profile"]
-
-
-def _loadJsonConfig(args, parser):
-    """
-    Load configuration from JSON file and merge with command line arguments.
-    CLI arguments take precedence over JSON values.
-
-    Args:
-        args: Parsed command line arguments
-        parser: The ArgumentParser instance
-
-    Returns:
-        argparse.Namespace: Updated arguments with JSON values
-    """
-    import json
-
-    cliProvidedArgs = set()
-    for arg in sys.argv[1:]:
-        if arg.startswith("--") and arg != "--json":
-            argName = arg[2:].replace("-", "_")
-            cliProvidedArgs.add(argName)
-
-    if len(cliProvidedArgs) > 0:
-        logAndPrint(
-            "Cannot use --json with other command line arguments. Use --json alone.",
-            "red",
-        )
-        sys.exit()
-
-    jsonPath = os.path.abspath(args.json)
-
-    if not os.path.exists(jsonPath):
-        logAndPrint(f"JSON config file not found: {jsonPath}", "red")
-        sys.exit()
-
-    try:
-        with open(jsonPath, "r", encoding="utf-8") as f:
-            jsonConfig = json.load(f)
-    except json.JSONDecodeError as e:
-        logAndPrint(f"Invalid JSON format in config file: {e}", "red")
-        sys.exit()
-    except Exception as e:
-        logAndPrint(f"Error reading JSON config: {e}", "red")
-        sys.exit()
-
-    defaults = {}
-    for action in parser._actions:
-        if action.dest not in ["help", "version", "json"]:
-            defaults[action.dest] = action.default
-
-    loadedKeys = set()
-    for key, value in jsonConfig.items():
-        if key == "json":
-            continue
-
-        if hasattr(args, key):
-            currentValue = getattr(args, key)
-            defaultValue = defaults.get(key)
-
-            if currentValue == defaultValue:
-                setattr(args, key, value)
-                logging.info(f"Loaded from JSON: {key} = {value}")
-            loadedKeys.add(key)
-        else:
-            logging.warning(f"Unknown option in JSON config: {key}")
-
-    return args, loadedKeys
-
-
-def capabilityMethods(parser, excluded=("decode_method",)):
-    """
-    Map capability -> its --*_method choices, harvested from the parser itself.
-
-    Single source of truth: the data comes straight from the registered actions,
-    so it can never drift from the actual argparse choices. 'decode_method'
-    (cpu/nvdec) is excluded since it's a backend toggle, not a model capability.
-    """
-    excluded = set(excluded)
-    capabilities = {}
-    for action in parser._actions:
-        if (
-            action.dest.endswith("_method")
-            and action.choices
-            and action.dest not in excluded
-        ):
-            capability = action.dest[: -len("_method")]
-            capabilities[capability] = [str(choice) for choice in action.choices]
-    return capabilities
-
-
-def _listMethods(parser, requested):
-    """
-    Print the --*_method choices for one capability (or all) and return an exit
-    code. Backend variants are grouped with the same logic as --help.
-    """
-    capabilities = capabilityMethods(parser)
-
-    palette = parser._palette() if hasattr(parser, "_palette") else {}
-    CYAN = palette.get("CYAN", "")
-    DIM = palette.get("DIM", "")
-    GREEN = palette.get("GREEN", "")
-    YELLOW = palette.get("YELLOW", "")
-    RED = palette.get("RED", "")
-    RESET = palette.get("RESET", "")
-
-    requested = (requested or "all").lower()
-
-    if requested != "all" and requested not in capabilities:
-        print(f"{RED}Unknown capability:{RESET} '{requested}'", file=sys.stderr)
-        suggestions = (
-            parser.getSuggestions(requested, list(capabilities))
-            if hasattr(parser, "getSuggestions")
-            else []
-        )
-        if suggestions:
-            print(
-                f"  {YELLOW}Did you mean:{RESET} "
-                f"{GREEN}{', '.join(suggestions)}{RESET}",
-                file=sys.stderr,
-            )
-        print(
-            f"  {DIM}Capabilities: {', '.join(sorted(capabilities))}{RESET}",
-            file=sys.stderr,
-        )
-        return 2
-
-    targets = (
-        capabilities
-        if requested == "all"
-        else {requested: capabilities[requested]}
-    )
-
-    for capability, methods in targets.items():
-        print(f"\n{CYAN}{capability}{RESET} {DIM}({len(methods)} methods){RESET}")
-        groups = TASHelpFormatter._group_choices(methods)
-        if len(groups) == 1:
-            print("  " + ", ".join(next(iter(groups.values()))))
-        else:
-            for backend, items in groups.items():
-                label = "%-10s" % ("[%s]" % backend)
-                print(f"  {DIM}{label}{RESET} {', '.join(items)}")
-
-    return 0
-
-
-def _buildParser(outputPath):
-    """
-    Build and configure the command line argument parser (without parsing).
-
-    Kept separate from createParser so the fully-wired parser can be introspected
-    (e.g. by --list_methods and the tests) without triggering a parse + validation.
-
-    Args:
-        outputPath (str): Default output directory path
-
-    Returns:
-        DidYouMeanArgumentParser: The configured parser.
-    """
-    argParser = DidYouMeanArgumentParser(
-        usage="main.py [options]",
-        formatter_class=TASHelpFormatter,
-    )
-
-    # Basic options
-    generalGroup = argParser.add_argument_group("General")
-    generalGroup.add_argument("-v", "--version", action="version", version=__version__)
-    generalGroup.add_argument("--input", type=str, help="Input video file")
-    generalGroup.add_argument("--output", type=str, help="Output video file")
-    generalGroup.add_argument(
-        "--inpoint", type=float, default=0, help="Input start time"
-    )
-    generalGroup.add_argument(
-        "--outpoint", type=float, default=0, help="Input end time"
-    )
-    generalGroup.add_argument(
-        "--preview", action="store_true", help="Preview the video during processing"
-    )
-    generalGroup.add_argument(
-        "--json",
-        type=str,
-        help="Path to JSON configuration file with processing options",
-    )
-
-    # Preset Configuration options
-    presetGroup = argParser.add_argument_group("Preset Configuration")
-    presetGroup.add_argument(
-        "--preset",
-        type=str,
-        help="Create and use a preset configuration file based on the current arguments",
-    )
-    presetGroup.add_argument(
-        "--list_presets", action="store_true", help="List all available presets"
-    )
-    presetGroup.add_argument(
-        "--list_methods",
-        type=str,
-        nargs="?",
-        const="all",
-        default=None,
-        metavar="CAPABILITY",
-        help="List available methods for a capability (e.g. --list_methods upscale) "
-        "or every capability when used alone, then exit.",
-    )
-
-    # Performance options
-    performanceGroup = argParser.add_argument_group("Performance")
-    performanceGroup.add_argument(
-        "--precision",
-        type=str,
-        choices=["fp32", "fp16"],
-        default="fp16",
-        help="NOT IMPLEMENTED YET! Precision for inference, default is fp16",
-    )
-    performanceGroup.add_argument(
-        "--decode_method",
-        type=str,
-        choices=["cpu", "nvdec"],
-        default="cpu",
-        help="Decoding backend to use, default is cpu. 'nvdec' requires an NVIDIA GPU with NVDEC support.",
-    )
-    performanceGroup.add_argument(
-        "--half",
-        type=str2bool,
-        nargs="?",
-        const=True,
-        default=True,
-        help="Use half precision for inference (default: True)",
-    )
-    performanceGroup.add_argument(
-        "--static",
-        action="store_true",
-        help="Force Static Mode engine generation for TensorRT",
-    )
-    performanceGroup.add_argument(
-        "--compile_mode",
-        type=str,
-        choices=["default", "max", "max-graphs"],
-        default="default",
-        help="[EXPERIMENTAL] Enable PyTorch compilation for CUDA models to improve performance. "
-        "Note: Only compatible with CUDA workflows and may cause compatibility issues with some models. "
-        "Increases startup time and memory usage. "
-        "'default' uses standard CudaGraph workflow without compilation, "
-        "'max' uses 'max-autotune-no-cudagraphs' mode, "
-        "'max-graphs' uses 'max-autotune-no-cudagraphs' with fullGraph=True. "
-        "Both 'max' options disable CudaGraphs, which may reduce performance at lower resolutions.",
-    )
-
-    # Interpolation options
-    _addInterpolationOptions(argParser)
-
-    # Upscaling options
-    _addUpscalingOptions(argParser)
-
-    # Deduplication options
-    _addDedupOptions(argParser)
-
-    # Video processing options
-    _addVideoProcessingOptions(argParser)
-
-    # Motion Blur options
-    _addMotionBlurOptions(argParser)
-
-    # Segmentation options
-    _addSegmentationOptions(argParser)
-
-    # Scene detection options
-    _addSceneDetectionOptions(argParser)
-
-    # Depth estimation options
-    _addDepthOptions(argParser)
-
-    # Encoding options
-    _addEncodingOptions(argParser)
-
-    # Object Detection options
-    objectGroup = argParser.add_argument_group("Object Detection")
-    objectGroup.add_argument(
-        "--obj_detect", action="store_true", help="Detect objects in the video"
-    )
-
-    objectGroup.add_argument(
-        "--obj_detect_method",
-        type=str,
-        default="yolov9_small-directml",
-        choices=[
-            "yolov9_small-directml",
-            "yolov9_medium-directml",
-            "yolov9_large-directml",
-            "yolov9_small-openvino",
-            "yolov9_medium-openvino",
-            "yolov9_large-openvino",
-            "yolov9_small-tensorrt",
-            "yolov9_medium-tensorrt",
-            "yolov9_large-tensorrt",
-        ],
-    )
-
-    objectGroup.add_argument(
-        "--obj_detect_disable_annotations",
-        type=str2bool,
-        default=False,
-        help="Disable class labels and confidence percentages on detection boxes (default: False)",
-    )
-
-    # Miscellaneous options
-    _addMiscOptions(argParser)
-
-    return argParser
-
-
-def createParser(outputPath):
-    """
-    Build the parser, parse argv, and run validation.
-
-    Args:
-        outputPath (str): Default output directory path
-
-    Returns:
-        argparse.Namespace: Parsed and validated command line arguments
-    """
-    argParser = _buildParser(outputPath)
-    args = argParser.parse_args()
-    return argumentsChecker(args, outputPath, argParser)
-
-
-def _addInterpolationOptions(argParser):
-    interpolationGroup = argParser.add_argument_group("Interpolation")
-    interpolationGroup.add_argument(
-        "--interpolate", action="store_true", help="Interpolate the video"
-    )
-    interpolationGroup.add_argument(
-        "--interpolate_factor", type=float, default=2, help="Interpolation factor"
-    )
-
-    interpolationMethods = [
-        "distildrba",
-        "distildrba-lite",
-        "distildrba-tensorrt",
-        "distildrba-lite-tensorrt",
-        "rife4.6",
-        "rife4.15-lite",
-        "rife4.16-lite",
-        "rife4.17",
-        "rife4.18",
-        "rife4.20",
-        "rife4.21",
-        "rife4.22",
-        "rife4.22-lite",
-        "rife4.25",
-        "rife4.25-lite",
-        "rife4.25-heavy",
-        "rife-ncnn",
-        "rife4.6-ncnn",
-        "rife4.15-lite-ncnn",
-        "rife4.16-lite-ncnn",
-        "rife4.17-ncnn",
-        "rife4.18-ncnn",
-        "rife4.20-ncnn",
-        "rife4.21-ncnn",
-        "rife4.22-ncnn",
-        "rife4.22-lite-ncnn",
-        "rife4.6-tensorrt",
-        "rife4.15-tensorrt",
-        "rife4.17-tensorrt",
-        "rife4.18-tensorrt",
-        "rife4.20-tensorrt",
-        "rife4.21-tensorrt",
-        "rife4.22-tensorrt",
-        "rife4.22-lite-tensorrt",
-        "rife4.25-tensorrt",
-        "rife4.25-lite-tensorrt",
-        "rife4.25-heavy-tensorrt",
-        "rife-tensorrt",
-        "gmfss",
-        "rife_elexor",
-        "rife_elexor-tensorrt",
-        "rife4.6-directml",
-        "rife4.15-directml",
-        "rife4.17-directml",
-        "rife4.18-directml",
-        "rife4.20-directml",
-        "rife4.21-directml",
-        "rife4.22-directml",
-        "rife4.22-lite-directml",
-        "rife4.25-directml",
-        "rife4.25-lite-directml",
-        "rife4.25-heavy-directml",
-        "rife4.6-openvino",
-        "rife4.15-openvino",
-        "rife4.17-openvino",
-        "rife4.18-openvino",
-        "rife4.20-openvino",
-        "rife4.21-openvino",
-        "rife4.22-openvino",
-        "rife4.22-lite-openvino",
-        "rife4.25-openvino",
-        "rife4.25-lite-openvino",
-        "rife4.25-heavy-openvino",
-        "rife-mps",
-        "rife4.6-mps",
-        "rife4.15-lite-mps",
-        "rife4.16-lite-mps",
-        "rife4.17-mps",
-        "rife4.18-mps",
-        "rife4.20-mps",
-        "rife4.21-mps",
-        "rife4.22-mps",
-        "rife4.22-lite-mps",
-        "rife4.25-mps",
-        "rife4.25-lite-mps",
-        "rife4.25-heavy-mps",
-        "rife_elexor-mps",
-    ]
-
-    interpolationGroup.add_argument(
-        "--interpolate_method",
-        type=str,
-        choices=interpolationMethods,
-        default="rife4.6",
-        help="Interpolation method",
-    )
-    interpolationGroup.add_argument(
-        "--slowmo",
-        action="store_true",
-        help="Enable slow motion interpolation, this will slow down the video instead of increasing the frame rate",
-    )
-    interpolationGroup.add_argument(
-        "--ensemble",
-        action="store_true",
-        help="Use the ensemble model for interpolation",
-    )
-    interpolationGroup.add_argument(
-        "--dynamic_scale",
-        action="store_true",
-        help="Use dynamic scaling for interpolation, this can improve the quality of the interpolation at the cost of performance, this is experimental and only works with Rife CUDA",
-    )
-    interpolationGroup.add_argument(
-        "--static_step",
-        action="store_true",
-        help="Force Static Timestep generation for Rife CUDA",
-    )
-    interpolationGroup.add_argument(
-        "--interpolate_first",
-        type=str2bool,
-        nargs="?",
-        const=True,
-        default=True,
-        choices=[True, False],
-        help="Switch back to the old approach where interpolated frames would instantly be written to the write queue",
-    )
-
-
-
-def _addUpscalingOptions(argParser):
-    upscaleGroup = argParser.add_argument_group("Upscaling")
-    upscaleGroup.add_argument(
-        "--upscale", action="store_true", help="Upscale the video"
-    )
-    upscaleGroup.add_argument(
-        "--upscale_factor",
-        type=int,
-        choices=[2, 3, 4],
-        default=2,
-        help="Upscaling factor (minimum 2)",
-    )
-
-    upscaleMethods = [
-        "shufflecugan",
-        "adore",
-        "adore-tensorrt",
-        "adore-directml",
-        "adore-openvino",
-        "fallin_soft",
-        "fallin_soft-tensorrt",
-        "fallin_soft-directml",
-        "fallin_strong",
-        "fallin_strong-tensorrt",
-        "fallin_strong-directml",
-        "span",
-        "shufflespan-directml",
-        "span-directml",
-        "shufflecugan-ncnn",
-        "shufflecugan-directml",
-        "shufflecugan-openvino",
-        "span-ncnn",
-        "span-tensorrt",
-        "shufflecugan-tensorrt",
-        "shufflespan-tensorrt",
-        "open-proteus",
-        "open-proteus-tensorrt",
-        "open-proteus-directml",
-        "aniscale2",
-        "aniscale2-tensorrt",
-        "aniscale2-directml",
-        "rtmosr",
-        "rtmosr-tensorrt",
-        "rtmosr-directml",
-        "saryn",
-        "saryn-tensorrt",
-        "saryn-directml",
-        "animesr",
-        "animesr-tensorrt",
-        "animesr-directml",
-        "animesr-openvino",
-        "span-openvino",
-        "open-proteus-openvino",
-        "aniscale2-openvino",
-        "shufflespan-openvino",
-        "rtmosr-openvino",
-        "saryn-openvino",
-        "fallin_soft-openvino",
-        "fallin_strong-openvino",
-        "gauss",
-        "gauss-tensorrt",
-        "gauss-directml",
-        "gauss-openvino",
-        "figsr",
-        "smosr",
-        "smosr-tensorrt",
-        "smosr-directml",
-        "smosr-openvino",
-        "artcnn_c4f16-tensorrt",
-        "artcnn_c4f16-directml",
-        "artcnn_c4f16-openvino",
-        "artcnn_c4f16_dn-tensorrt",
-        "artcnn_c4f16_dn-directml",
-        "artcnn_c4f16_dn-openvino",
-        "artcnn_c4f16_ds-tensorrt",
-        "artcnn_c4f16_ds-directml",
-        "artcnn_c4f16_ds-openvino",
-        "artcnn_c4f32-tensorrt",
-        "artcnn_c4f32-directml",
-        "artcnn_c4f32-openvino",
-        "artcnn_c4f32_dn-tensorrt",
-        "artcnn_c4f32_dn-directml",
-        "artcnn_c4f32_dn-openvino",
-        "artcnn_c4f32_ds-tensorrt",
-        "artcnn_c4f32_ds-directml",
-        "artcnn_c4f32_ds-openvino",
-        "artcnn_r8f64-tensorrt",
-        "artcnn_r8f64-directml",
-        "artcnn_r8f64-openvino",
-        "artcnn_r16f96-tensorrt",
-        "artcnn_r16f96-directml",
-        "artcnn_r16f96-openvino",
-        "shufflecugan-mps",
-        "adore-mps",
-        "span-mps",
-        "open-proteus-mps",
-        "aniscale2-mps",
-        "shufflespan-mps",
-        "rtmosr-mps",
-        "saryn-mps",
-        "fallin_soft-mps",
-        "fallin_strong-mps",
-        "gauss-mps",
-        "figsr-mps",
-        "smosr-mps",
-        "maxine-bicubic",
-        "maxine-low",
-        "maxine-medium",
-        "maxine-high",
-        "maxine-ultra",
-        "maxine-highbitrate_low",
-        "maxine-highbitrate_medium",
-        "maxine-highbitrate_high",
-        "maxine-highbitrate_ultra",
-    ]
-
-    upscaleGroup.add_argument(
-        "--upscale_method",
-        type=str,
-        choices=upscaleMethods,
-        default="shufflecugan",
-        help="Upscaling method",
-    )
-    upscaleGroup.add_argument(
-        "--custom_model",
-        type=str,
-        default="",
-        help="Path to a custom upscale model. Use .pt/.pth/.ckpt/.safetensors with CUDA methods and .onnx with -directml, -openvino, or -tensorrt methods",
-    )
-
-
-def _addDedupOptions(argParser):
-    dedupGroup = argParser.add_argument_group("Deduplication")
-    dedupGroup.add_argument(
-        "--dedup", action="store_true", help="Deduplicate the video"
-    )
-    dedupGroup.add_argument(
-        "--dedup_method",
-        type=str,
-        default="ssim",
-        choices=[
-            "ssim",
-            "mse",
-            "ssim-cuda",
-            "mse-cuda",
-            "flownets",
-            "vmaf",
-            "vmaf-cuda",
-        ],
-        help="Deduplication method",
-    )
-    dedupGroup.add_argument(
-        "--dedup_sens", type=float, default=35, help="Deduplication sensitivity"
-    )
-    dedupGroup.add_argument(
-        "--smooth_dedup",
-        action="store_true",
-        help="Smooth deduplication, this will remove duplicates while also generating new frames to make the video smoother, this is experimental and may not work well with all videos, use --interpolate_method to set the interpolation method",
-    )
-
-
-def _addVideoProcessingOptions(argParser):
-    processingGroup = argParser.add_argument_group("Video Processing")
-    processingGroup.add_argument(
-        "--restore", action="store_true", help="Restore the video"
-    )
-    processingGroup.add_argument(
-        "--stabilize", action="store_true", help="Stabilize the video"
-    )
-
-    restoreMethods = [
-        "scunet",
-        "scunet-tensorrt",
-        "scunet-directml",
-        "scunet-openvino",
-        "nafnet",
-        "dpir",
-        "real-plksr",
-        "anime1080fixer",
-        "anime1080fixer-tensorrt",
-        "anime1080fixer-directml",
-        "anime1080fixer-openvino",
-        "fastlinedarken",
-        "fastlinedarken-tensorrt",
-        "autocas",
-        "gater3",
-        "gater3-directml",
-        "gater3-openvino",
-        "deepdeband-f",
-        "deh264_real",
-        "deh264_real-tensorrt",
-        "deh264_real-directml",
-        "deh264_real-openvino",
-        "deh264_span",
-        "deh264_span-tensorrt",
-        "deh264_span-directml",
-        "deh264_span-openvino",
-        "hurrdeblur",
-        "hurrdeblur-tensorrt",
-        "hurrdeblur-directml",
-        "hurrdeblur-openvino",
-        "dehalo",
-        "dehalo-tensorrt",
-        "dehalo-directml",
-        "dehalo-openvino",
-        "scunet-mps",
-        "nafnet-mps",
-        "dpir-mps",
-        "real-plksr-mps",
-        "anime1080fixer-mps",
-        "gater3-mps",
-        "deh264_real-mps",
-        "deh264_span-mps",
-        "hurrdeblur-mps",
-        "dehalo-mps",
-        "linethinner-lite",
-        "linethinner-medium",
-        "linethinner-heavy",
-        "linethinner-lite-cuda",
-        "linethinner-medium-cuda",
-        "linethinner-heavy-cuda",
-        "maxine-denoise_low",
-        "maxine-denoise_medium",
-        "maxine-denoise_high",
-        "maxine-denoise_ultra",
-        "maxine-deblur_low",
-        "maxine-deblur_medium",
-        "maxine-deblur_high",
-        "maxine-deblur_ultra",
-    ]
-
-    processingGroup.add_argument(
-        "--restore_method",
-        type=str,
-        nargs="+",
-        default=["anime1080fixer"],
-        choices=restoreMethods,
-        help="Denoising method(s), can specify multiple for chaining",
-    )
-    processingGroup.add_argument(
-        "--resize", action="store_true", help="Resize the video"
-    )
-    processingGroup.add_argument(
-        "--resize_factor",
-        type=float,
-        default=2,
-        help="Resize factor (can be between 0 and 1 for downscaling)",
-    )
-    processingGroup.add_argument(
-        "--output_scale",
-        type=str,
-        default="",
-        help="Output resolution in WIDTHxHEIGHT format (e.g., 2560x1440)",
-    )
-
-
-def _addMotionBlurOptions(argParser):
-    moblurGroup = argParser.add_argument_group("Motion Blur")
-    moblurGroup.add_argument(
-        "--moblur",
-        action="store_true",
-        help="Add motion blur via interpolation and frame blending",
-    )
-    moblurGroup.add_argument(
-        "--moblur_method",
-        type=str,
-        default="rife4.25",
-        choices=[
-            "rife4.6",
-            "rife4.25",
-            "rife4.6-directml",
-            "rife4.25-directml",
-            "rife4.6-tensorrt",
-            "rife4.25-tensorrt",
-        ],
-        help="Interpolation model + backend used to synthesize the motion-blur samples. "
-        "rife4.25 is sharper/slower, rife4.6 is faster. Suffix picks the backend: "
-        "none = CUDA (NVIDIA; the only path that gets the windowed-sample speedup), "
-        "-tensorrt = CUDA via TensorRT, -directml = AMD/Intel GPUs.",
-    )
-    moblurGroup.add_argument(
-        "--moblur_factor",
-        type=int,
-        default=8,
-        help="Interpolation factor for motion blur sample generation (8-16 recommended; higher = smoother trails)",
-    )
-    moblurGroup.add_argument(
-        "--moblur_strength",
-        type=str,
-        default="gaussian_sym",
-        choices=["equal", "gaussian_sym", "pyramid", "ascending", "descending"],
-        help="Frame-weighting scheme. gaussian_sym mimics natural shutter falloff",
-    )
-    moblurGroup.add_argument(
-        "--moblur_shutter_angle",
-        type=float,
-        default=180.0,
-        help="Virtual shutter angle in degrees (0-360). 180 = cinema standard, 360 = max smear, 90 = crisp, 0 = disabled",
-    )
-    moblurGroup.add_argument(
-        "--moblur_no_linear_blend",
-        action="store_true",
-        help="Disable gamma-correct (linear-light) blending. Not recommended — produces muddy highlights and lifted shadows",
-    )
-    moblurGroup.add_argument(
-        "--moblur_mask",
-        type=str,
-        default="",
-        help="Path to a transparent PNG marking regions to protect from motion blur. Paint opaque dark pixels on a transparent background — those areas stay sharp (HUD/text), transparent areas blur normally.",
-    )
-
-
-def _addSegmentationOptions(argParser):
-    segmentationGroup = argParser.add_argument_group(
-        "Segmentation / Background Removal"
-    )
-    segmentationGroup.add_argument(
-        "--segment", action="store_true", help="Segment the video"
-    )
-    segmentationGroup.add_argument(
-        "--segment_method",
-        type=str,
-        default="anime",
-        choices=["anime", "anime-tensorrt", "anime-directml", "cartoon"],
-        help="Segmentation method",
-    )
-
-
-def _addSceneDetectionOptions(argParser):
-    sceneGroup = argParser.add_argument_group("Scene Detection")
-    sceneGroup.add_argument(
-        "--autoclip", action="store_true", help="Detect scene changes"
-    )
-    sceneGroup.add_argument(
-        "--autoclip_method",
-        type=str,
-        default="pyscenedetect",
-        choices=[
-            "pyscenedetect",
-            "maxxvit-directml",
-            "maxxvit-tensorrt",
-            "transnetv2",
-        ],
-        help="Autoclip detection backend",
-    )
-    sceneGroup.add_argument(
-        "--autoclip_sens", type=float, default=50, help="Autoclip sensitivity"
-    )
-
-
-def _addDepthOptions(argParser):
-    depthGroup = argParser.add_argument_group("Depth Estimation")
-    depthGroup.add_argument(
-        "--depth", action="store_true", help="Estimate the depth of the video"
-    )
-
-    depthMethods = [
-        "small_v2",
-        "base_v2",
-        "large_v2",
-        "giant_v2",
-        "distill_small_v2",
-        "distill_base_v2",
-        "distill_large_v2",
-        "og_small_v2",
-        "og_base_v2",
-        "og_large_v2",
-        "og_giant_v2",
-        "og_distill_small_v2",
-        "og_distill_base_v2",
-        "og_distill_large_v2",
-        "og_video_small_v2",
-        "og_video_base_v2",
-        "og_video_large_v2",
-        "video_small_v2",
-        "video_large_v2",
-        "small_v2-tensorrt",
-        "base_v2-tensorrt",
-        "large_v2-tensorrt",
-        "distill_small_v2-tensorrt",
-        "distill_base_v2-tensorrt",
-        "distill_large_v2-tensorrt",
-        "small_v2-directml",
-        "base_v2-directml",
-        "large_v2-directml",
-        "distill_small_v2-directml",
-        "distill_base_v2-directml",
-        "distill_large_v2-directml",
-        "og_small_v2-tensorrt",
-        "og_base_v2-tensorrt",
-        "og_large_v2-tensorrt",
-        "og_distill_small_v2-tensorrt",
-        "og_distill_base_v2-tensorrt",
-        "og_distill_large_v2-tensorrt",
-        "small_v3",
-        "base_v3",
-        "og_large_v3",
-        "small_v3-directml",
-        "base_v3-directml",
-        "small_v3-tensorrt",
-        "base_v3-tensorrt",
-        "small_v2-openvino",
-        "base_v2-openvino",
-        "large_v2-openvino",
-        "distill_small_v2-openvino",
-        "distill_base_v2-openvino",
-        "distill_large_v2-openvino",
-        "og_small_v2-openvino",
-        "og_base_v2-openvino",
-        "og_large_v2-openvino",
-        "small_v3-openvino",
-        "base_v3-openvino",
-    ]
-
-    depthGroup.add_argument(
-        "--depth_method",
-        type=str,
-        choices=depthMethods,
-        default="small_v2",
-        help="Depth estimation method",
-    )
-    depthGroup.add_argument(
-        "--depth_quality",
-        type=str,
-        choices=["low", "medium", "high"],
-        default="low",
-        help="This will determine the quality of the depth map, low is significantly faster but lower quality, only works with CUDA Depth Maps",
-    )
-    depthGroup.add_argument(
-        "--depth_norm",
-        action="store_true",
-        help="Apply sliding window normalization to reduce flickering in depth maps, not compatible with video depth methods",
-    )
-
-
-def _addEncodingOptions(argParser):
-    encodingGroup = argParser.add_argument_group("Encoding")
-
-    encodeMethods = [
-        "x264",
-        "slow_x264",
-        "x264_10bit",
-        "x264_animation",
-        "x264_animation_10bit",
-        "x265",
-        "slow_x265",
-        "x265_10bit",
-        "nvenc_h264",
-        "slow_nvenc_h264",
-        "nvenc_h265",
-        "slow_nvenc_h265",
-        "nvenc_h265_10bit",
-        "nvenc_av1",
-        "slow_nvenc_av1",
-        "qsv_h264",
-        "qsv_h265",
-        "qsv_h265_10bit",
-        "av1",
-        "slow_av1",
-        "h264_amf",
-        "hevc_amf",
-        "hevc_amf_10bit",
-        "prores",
-        "prores_segment",
-        "gif",
-        "vp9",
-        "qsv_vp9",
-        "lossless",
-        "lossless_nvenc",
-        "png",
-        "nvenc_h264_nelux",
-        "nvenc_h265_nelux",
-        "nvenc_av1_nelux",
-        "x264_nelux",
-        "x265_nelux",
-        "av1_nelux",
-    ]
-
-    encodingGroup.add_argument(
-        "--encode_method",
-        type=str,
-        choices=encodeMethods,
-        default="x264",
-        help="Encoding method",
-    )
-    encodingGroup.add_argument(
-        "--custom_encoder", type=str, default="", help="Custom encoder settings"
-    )
-
-
-def _addMiscOptions(argParser):
-    miscGroup = argParser.add_argument_group("Miscellaneous")
-    miscGroup.add_argument(
-        "--benchmark", action="store_true", help="Benchmark the script"
-    )
-    miscGroup.add_argument(
-        "--profile",
-        action="store_true",
-        help="Enable torch.profiler to analyze GPU/CPU performance bottlenecks",
-    )
-    miscGroup.add_argument(
-        "--offline",
-        type=str,
-        nargs="*",
-        default="none",
-        help="Download a specific model or multiple models for offline use, use keyword 'all' to download all models",
-    )
-    miscGroup.add_argument(
-        "--ae",
-        type=str,
-        default="",
-        help="Notify if script is run from After Effects interface",
-    )
-    miscGroup.add_argument(
-        "--bit_depth",
-        type=str,
-        default="8bit",
-        choices=["8bit", "16bit"],
-        help="Bit Depth of the raw pipe input to FFMPEG",
-    )
-    miscGroup.add_argument(
-        "--download_requirements",
-        type=str,
-        nargs="?",
-        const="",
-        default=None,
-        metavar="PROFILE",
-        help=(
-            "Download dependencies for a runtime profile. Supported profiles: "
-            "windows-cuda, windows-lite, linux-cuda, linux-lite. When used "
-            "without a profile, prompts for the current OS full CUDA / TensorRT "
-            "or lite dependencies, with guidance for newer NVIDIA GPUs vs "
-            "GTX 10xx, AMD, and Intel hardware."
-        ),
-    )
-    miscGroup.add_argument(
-        "--cleanup",
-        action="store_true",
-        help="Remove unused libraries from the script, in case if there were migrations or changes in the script",
-    )
-
-
-def _autoEnableParentFlags(args):
-    """
-    Automatically enable parent flags when *_method arguments are provided.
-    For example: --upscale_method shufflecugan will auto-enable --upscale
-
-    Args:
-        args: Parsed command line arguments
-    """
-    logging.info(f"[DEBUG] hasattr(args, '_json_keys'): {hasattr(args, '_json_keys')}")
-    if hasattr(args, "_json_keys"):
-        logging.info(f"[DEBUG] args._json_keys: {args._json_keys}")
-
-    methodToFlagMapping = {
-        "interpolate_method": ("interpolate", "rife4.6"),
-        "interpolate_factor": ("interpolate", 2.0),
-        "upscale_method": ("upscale", "shufflecugan"),
-        "upscale_factor": ("upscale", 2),
-        "dedup_method": ("dedup", "ssim"),
-        "dedup_sens": ("dedup", 35.0),
-        "restore_method": ("restore", ["anime1080fixer"]),
-        "segment_method": ("segment", "anime"),
-        "depth_method": ("depth", "small_v2"),
-        "obj_detect_method": ("obj_detect", "yolov9_small-directml"),
-        "resize_factor": ("resize", 2),
-        "output_scale": ("resize", ""),
-        "moblur_method": ("moblur", "rife4.25"),
-        "moblur_factor": ("moblur", 8),
-        "moblur_strength": ("moblur", "gaussian_sym"),
-        "moblur_shutter_angle": ("moblur", 180.0),
-    }
-
-    cliProvided = set()
-    for a in sys.argv[1:]:
-        if a.startswith("--"):
-            name = a[2:].split("=")[0]
-            cliProvided.add(name)
-
-    for methodArg, (parentFlag, defaultValue) in methodToFlagMapping.items():
-        if hasattr(args, methodArg):
-            currentValue = getattr(args, methodArg)
-
-            providedOnCLI = (
-                methodArg in cliProvided or methodArg.replace("_", "-") in cliProvided
-            )
-
-            isExplicitlyProvided = providedOnCLI or (
-                hasattr(args, "_json_keys") and methodArg in args._json_keys
-            )
-
-            if methodArg == "interpolate_method":
-                logging.info(
-                    f"[DEBUG] interpolate_method - providedOnCLI: {providedOnCLI}, isExplicitlyProvided: {isExplicitlyProvided}"
-                )
-
-            if isExplicitlyProvided:
-                if not getattr(args, parentFlag):
-                    setattr(args, parentFlag, True)
-                    logging.info(
-                        f"Auto-enabling --{parentFlag} because --{methodArg} was provided"
-                    )
-            else:
-                if currentValue != defaultValue:
-                    if not getattr(args, parentFlag):
-                        setattr(args, parentFlag, True)
-                        logging.info(
-                            f"Auto-enabling --{parentFlag} because {methodArg} differs from default"
-                        )
-
-
-def argumentsChecker(args, outputPath, parser):
-    args.png_passthrough = False
-    args.single_image_input = False
-
-    if args.list_presets:
-        from src.utils.presetLogic import listPresets
-
-        listPresets()
-        sys.exit()
-
-    if args.list_methods is not None:
-        sys.exit(_listMethods(parser, args.list_methods))
-
-    if args.preset:
-        from src.utils.presetLogic import createPreset
-
-        args = createPreset(args)
-
-    if args.json:
-        args, loadedKeys = _loadJsonConfig(args, parser)
-        args._json_keys = loadedKeys
-
-    _autoEnableParentFlags(args)
-
-    if args.download_requirements is not None:
-        from src.utils.dependencyHandler import DependencyChecker
-
-        _handleDependencies(args)
-        selectedProfile = args.download_requirements.strip().lower()
-        if not selectedProfile:
-            selectedProfile = _promptDownloadRequirementsSelection()
-
-        checker = DependencyChecker()
-        if not checker.installProfile(selectedProfile):
-            sys.exit(1)
-
-        logAndPrint(
-            "All required libraries have been downloaded, you can now run the script freely.",
-            "green",
-        )
-        sys.exit()
-
-    if args.cleanup:
-        from src.utils.dependencyHandler import (
-            DependencyChecker,
-            uninstallDependencies,
-            getRequirementsFileForProfile,
-            getDependencyProfile,
-        )
-
-        checker = DependencyChecker()
-        storedProfile = checker.loadStoredProfile()
-
-        if storedProfile:
-            try:
-                extension = getRequirementsFileForProfile(storedProfile)
-            except ValueError:
-                storedProfile = None
-
-        if not storedProfile:
-            from src.utils.isCudaInit import detectNVidiaGPU, detectGPUArchitecture
-
-            isNvidia = detectNVidiaGPU()
-            supportsCuda = False
-            if isNvidia:
-                supportsCuda, _, _ = detectGPUArchitecture()
-            extension = getRequirementsFileForProfile(
-                getDependencyProfile(cs.SYSTEM, supportsCuda)
-            )
-
-        success, message = uninstallDependencies(extension=extension)
-        checker.clearCache()
-
-        logging.info(message)
-
-        if success:
-            logAndPrint(
-                "Dependencies from the selected runtime profile were uninstalled.",
-                "green",
-            )
-            sys.exit()
-        else:
-            logAndPrint(
-                "Failed to uninstall dependencies, please check the logs for more details",
-                "red",
-            )
-            logging.error("Failed to uninstall dependencies")
-            print(message)
-
-    logging.info("============== Version ==============")
-    logging.info(f"TAS: {__version__}\n")
-
-    cliArgs = sys.argv[1:]
-    providedArgs = set()
-    for i, arg in enumerate(cliArgs):
-        if arg.startswith("--"):
-            argName = arg[2:]
-            providedArgs.add(argName)
-
-    logging.info("============== Arguments ==============")
-    for arg, value in vars(args).items():
-        if arg in providedArgs and value not in [None, "", "none"]:
-            logging.info(f"{arg.upper()}: {value}")
-
-    if not args.benchmark:
-        from .checkSpecs import checkSystem
-
-        checkSystem()
-
-    if args.preview and args.benchmark:
-        logAndPrint(
-            "Preview cannot be enabled in benchmark mode, disabling preview",
-            "yellow",
-        )
-        args.preview = False
-
-    if args.ae:
-        logging.info("After Effects interface detected")
-        cs.ADOBE = True
-        from src.utils.aeComms import startServerInThread
-
-        try:
-            startServerInThread(
-                host=args.ae,
-            )
-        except Exception as e:
-            logging.error(f"Failed to start AE comms server: {e}")
-            logAndPrint(
-                "Failed to start AE comms server, please check the logs for more details",
-                "red",
-            )
-
-    logging.info("\n============== Arguments Checker ==============")
-    _handleDependencies(args)
-
-    if args.slowmo and not args.interpolate:
-        logAndPrint(
-            "Slow motion is enabled but interpolation is not, disabling slowmo",
-            "yellow",
-        )
-        args.slowmo = False
-
-    _handleDepthSettings(args)
-
-    if args.offline != "none":
-        _downloadOfflineModels(args)
-
-    _configureProcessingSettings(args)
-
-    # Check CUDA availability and adjust methods if needed
-    _adjustMethodsBasedOnCuda(args)
-    _validateCustomUpscaleModel(args)
-
-    if args.custom_encoder:
-        logging.info("Custom encoder specified, use with caution")
-
-    if args.bit_depth == "16bit" and args.segment:
-        logging.error(
-            "16bit input is not supported with segmentation, defaulting to 8bit"
-        )
-        args.bit_depth = "8bit"
-
-    if args.output:
-        outDir = os.path.dirname(os.path.abspath(args.output))
-        if outDir and not os.path.exists(outDir):
-            os.makedirs(outDir, exist_ok=True)
-
-    if args.encode_method in ["gif", "png"]:
-        logging.info(
-            f"Encoding method is set to {args.encode_method}, disabling audio processing"
-        )
-        cs.AUDIO = False
-
-    if not args.input:
-        logging.error("No input specified")
-        sys.exit()
-    elif args.input.startswith(("http", "www")):
-        processURL(args, outputPath)
-    elif args.input.lower().endswith((".png", ".jpg", ".jpeg")):
-        if "%" in args.input:
-            logging.info(f"Image sequence pattern detected: {args.input}")
-            args.input = os.path.abspath(args.input)
-            cs.AUDIO = False
-        elif args.input.lower().endswith(".png"):
-            args.input = os.path.abspath(args.input)
-            cs.AUDIO = False
-            args.single_image_input = True
-            args.png_passthrough = True
-            logging.info("Single PNG input detected, enabling PNG passthrough mode")
-
-            if isAnyOtherProcessingMethodEnabled(args) and args.encode_method != "png":
-                logging.info(
-                    "Single PNG with processing detected; forcing --encode_method png for valid image output"
-                )
-                args.encode_method = "png"
-        else:
-            raise Exception(
-                "Single image input is not supported for this format. For image sequences, use a pattern like 'frames_%05d.png' or provide a folder containing PNG files."
-            )
-    elif args.input.lower().endswith((".gif")):
-        if args.encode_method != "gif":
-            logging.error(
-                "GIF input detected but encoding method is not set to GIF, defaulting to GIF encoding"
-            )
-            args.encode_method = "gif"
-    else:
-        try:
-            args.input = os.path.abspath(args.input)
-        except Exception:
-            logging.error("Error processing input")
-            sys.exit()
-
-    if args.output_scale:
-        try:
-            width, height = args.output_scale.split("x")
-            args.output_scale_width = int(width)
-            args.output_scale_height = int(height)
-            if args.output_scale_width <= 0 or args.output_scale_height <= 0:
-                raise ValueError("Width and height must be positive integers")
-            logging.info(
-                f"Output scale set to {args.output_scale_width}x{args.output_scale_height}"
-            )
-        except (ValueError, AttributeError):
-            logAndPrint(
-                f"Invalid output_scale format: {args.output_scale}. Expected format: WIDTHxHEIGHT (e.g., 2560x1440)"
-            )
-            sys.exit()
-    else:
-        args.output_scale_width = None
-        args.output_scale_height = None
-
-    if args.upscale and hasattr(args, "upscale_factor"):
-        try:
-            if int(args.upscale_factor) < 2:
-                logAndPrint(
-                    "Upscale factor must be at least 2 when --upscale is enabled; defaulting to 2",
-                    "yellow",
-                )
-                logging.info(
-                    f"Adjusted upscale_factor from {args.upscale_factor} to 2 to satisfy minimum requirement"
-                )
-                args.upscale_factor = 2
-        except Exception:
-            logAndPrint(
-                "Invalid upscale_factor provided; defaulting to 2",
-                "yellow",
-            )
-            logging.info("Invalid upscale_factor value encountered; set to 2")
-            args.upscale_factor = 2
-
-    logging.info(
-        f"[DEBUG] Before processing check - args.interpolate: {args.interpolate}"
-    )
-
-    if not isAnyOtherProcessingMethodEnabled(args) and not args.png_passthrough:
-        logAndPrint(
-            "No processing methods specified, make sure to use enabler arguments like --upscale, --interpolate, etc.",
-            "red",
-        )
-        sys.exit()
-
-    return args
-
-
-def _handleDependencies(args):
-    legacyFFMPEG = os.path.join(cs.WHEREAMIRUNFROM, "ffmpeg")
-    if os.path.isdir(legacyFFMPEG):
-        try:
-            from src.utils.getFFMPEG import remove_readonly
-
-            shutil.rmtree(legacyFFMPEG, onerror=remove_readonly)
-            logging.info(f"Removed legacy FFmpeg folder: {legacyFFMPEG}")
-        except Exception as e:
-            logging.warning(f"Failed to remove legacy FFmpeg folder: {e}")
-
-    ffmpegBaseDir = cs.WHEREAMIRUNFROM
-    ffmpegSharedDir = os.path.join(ffmpegBaseDir, "ffmpeg_shared")
-
-    cs.FFMPEGPATH = os.path.join(
-        ffmpegSharedDir,
-        "ffmpeg.exe" if cs.SYSTEM == "Windows" else "ffmpeg",
-    )
-
-    cs.FFPROBEPATH = os.path.join(
-        ffmpegSharedDir,
-        "ffprobe.exe" if cs.SYSTEM == "Windows" else "ffprobe",
-    )
-
-    if cs.SYSTEM == "Windows":
-        if "ffprobe.exe" not in os.environ["PATH"]:
-            os.environ["PATH"] += os.pathsep + os.path.dirname(cs.FFPROBEPATH)
-    else:
-        if "ffprobe" not in os.environ["PATH"]:
-            os.environ["PATH"] += os.pathsep + os.path.dirname(cs.FFPROBEPATH)
-
-    if not os.path.exists(cs.FFMPEGPATH) or not os.path.exists(cs.FFPROBEPATH):
-        from src.utils.getFFMPEG import getFFMPEG
-
-        getFFMPEG()
-
-    if cs.SYSTEM == "Windows":
-        ffmpeg_dir = os.path.dirname(cs.FFMPEGPATH)
-        if os.path.exists(ffmpeg_dir):
-            try:
-                os.add_dll_directory(ffmpeg_dir)
-                logging.info(f"Added FFmpeg directory to DLL search path: {ffmpeg_dir}")
-            except Exception as e:
-                logging.warning(f"Failed to add FFmpeg to DLL search path: {e}")
-
-    try:
-        from src.utils.isCudaInit import detectNVidiaGPU, detectGPUArchitecture
-
-        isNvidia = detectNVidiaGPU()
-        supportsCuda = False
-        if isNvidia:
-            supportsCuda, _, _ = detectGPUArchitecture()
-        args.supportsCuda = supportsCuda
-    except ImportError:
-        isNvidia = False
-        supportsCuda = False
-        args.supportsCuda = False
-
-    from src.utils.dependencyHandler import getDependencyProfile
-
-    args.dependency_profile = getDependencyProfile(cs.SYSTEM, supportsCuda)
-
-    if args.download_requirements is None and not args.cleanup:
-        from src.utils.dependencyHandler import DependencyChecker
-
-        checker = DependencyChecker()
-        checker.ensureDependencies()
-
-
-def _handleDepthSettings(args):
-    if args.depth:
-        logging.info("Depth enabled, audio processing will be disabled")
-        cs.AUDIO = False
-
-    if args.depth_quality not in ["low"] and args.depth_method.split("-")[-1] in [
-        "tensorrt",
-        "directml",
-    ]:
-        logAndPrint(
-            f"{args.depth_quality.upper()} depth estimation quality is incomaptible with tensorrt and directml, defaulting to low quality",
-            "yellow",
-        )
-        args.depth_quality = "low"
-
-    if args.depth_norm and "video" in args.depth_method:
-        logAndPrint(
-            "Depth normalization is not compatible with video depth methods, disabling depth_norm",
-            "yellow",
-        )
-        args.depth_norm = False
-
-    if args.depth_method in ["giant_v2", "og_giant_v2"]:
-        logAndPrint(
-            f"{args.depth_method} is a very large model and may cause out of memory errors on GPUs with less than 16GB of VRAM",
-            "yellow",
-        )
-        if args.half:
-            logAndPrint(
-                "Half precision does not work with giant models, switching to full precision",
-            )
-            args.half = False
-
-
-def _downloadOfflineModels(args):
-    from .downloadModels import downloadModels, modelsList
-
-    logAndPrint(
-        f"Offline mode enabled, downloading {args.offline} model(s)...", "green"
-    )
-
-    options = modelsList() if args.offline == ["all"] else args.offline
-    for option in options:
-        for precision in [True, False]:
-            try:
-                downloadModels(option.lower(), half=precision)
-            except Exception:
-                logging.error(
-                    f"Failed to download model: {option} with precision: {'fp16' if precision else 'fp32'}"
-                )
-    logAndPrint(
-        f"Offline model(s) {', '.join(options)} downloaded successfully!", "green"
-    )
-
-
-def _configureProcessingSettings(args):
-    if args.slowmo:
-        cs.AUDIO = False
-        logging.info("Slow motion enabled, audio processing disabled")
-
-    if args.static_step and isinstance(args.interpolate_factor, float):
-        logging.info("Interpolate Factor is a float, static step will be disabled")
-        args.static_step = False
-
-    if args.dedup:
-        if not args.smooth_dedup:
-            cs.AUDIO = False
-            logging.info(
-                "Deduplication enabled and smooth dedup disabled, audio processing disabled"
-            )
-
-        if args.dedup_method in ["ssim", "ssim-cuda"]:
-            args.dedup_sens = 1.0 - (args.dedup_sens / 1000)
-        elif args.dedup_method in ["flownets"]:
-            args.dedup_sens = args.dedup_sens / 100
-
-        logging.info(
-            f"New dedup sensitivity for {args.dedup_method} is: {args.dedup_sens}"
-        )
-
-    if args.autoclip:
-        if args.autoclip_method == "pyscenedetect":
-            # AdaptiveDetector treats higher threshold as less sensitive; flip
-            # so the user-facing sens is monotonic (higher = more cuts).
-            args.autoclip_sens = float(100 - args.autoclip_sens)
-        else:
-            # maxxvit / transnetv2 produce probability in [0, 1]; map sens
-            # 0..100 -> threshold 1..0 so higher sens fires more cuts.
-            args.autoclip_sens = float(1.0 - (args.autoclip_sens / 100.0))
-        logging.info(
-            f"New autoclip sensitivity for {args.autoclip_method} is: {args.autoclip_sens}"
-        )
-
-    if args.compile_mode != "default":
-        logging.info(
-            f"Pytorch Compile mode is set to {args.compile_mode}, this will increase startup time and memory usage and may lead to instability with some models"
-        )
-
-
-def _validateCustomUpscaleModel(args):
-    if not args.custom_model:
-        return
-
-    args.custom_model = os.path.abspath(args.custom_model)
-    if not os.path.isfile(args.custom_model):
-        logAndPrint(f"Custom model file not found: {args.custom_model}", "red")
-        sys.exit()
-
-    extension = os.path.splitext(args.custom_model)[1].lower()
-    pytorchExtensions = {".pt", ".pth", ".ckpt", ".safetensors"}
-    onnxExtensions = {".onnx"}
-    backendSuffixes = ("-directml", "-openvino", "-tensorrt", "-ncnn")
-
-    selectedBackend = "pytorch"
-    baseMethod = args.upscale_method
-    for suffix in backendSuffixes:
-        if args.upscale_method.endswith(suffix):
-            selectedBackend = suffix[1:]
-            baseMethod = args.upscale_method[: -len(suffix)]
-            break
-
-    if extension in onnxExtensions:
-        if selectedBackend not in {"directml", "openvino", "tensorrt"}:
-            logAndPrint(
-                "Custom ONNX upscale models require an ONNX backend. Use an upscale method ending in -directml, -openvino, or -tensorrt, for example "
-                f"{baseMethod}-directml.",
-                "red",
-            )
-            sys.exit()
-        return
-
-    if extension in pytorchExtensions:
-        if selectedBackend != "pytorch":
-            logAndPrint(
-                "Custom PyTorch upscale models require a CUDA/PyTorch upscale method without a backend suffix. "
-                f"Use {baseMethod} for .pt/.pth/.ckpt/.safetensors files.",
-                "red",
-            )
-            sys.exit()
-        return
-
-    logAndPrint(
-        "Unsupported custom upscale model format. Supported extensions are .pt, .pth, .ckpt, .safetensors, and .onnx.",
-        "red",
-    )
-    sys.exit()
-
-
-def _adjustMethodsBasedOnCuda(args):
-    isDarwin = cs.SYSTEM == "Darwin"
-
-    def adjustMethod(method, modelsList):
-        base = method.lower()
-        if isDarwin:
-            mpsMethod = f"{base}-mps"
-            if mpsMethod in modelsList:
-                return mpsMethod
-        directML = f"{base}-directml"
-        if directML in modelsList:
-            return directML
-        newMethod = f"{base}-ncnn"
-        if newMethod in modelsList:
-            return newMethod
-        return method
-
-    supportsCuda = getattr(args, "supportsCuda", None)
-
-    if supportsCuda is None:
-        from src.utils.isCudaInit import CudaChecker, detectGPUArchitecture
-
-        isCuda = CudaChecker()
-
-        needsFallback = False
-        if isCuda.cudaAvailable:
-            isModernGPU, gpuName, computeCap = detectGPUArchitecture()
-            if not isModernGPU:
-                logAndPrint(
-                    f"Detected {gpuName} (compute capability: {computeCap}). "
-                    f"This GPU may not support modern CUDA kernels. "
-                    f"Automatically switching to DirectML/NCNN backends for compatibility.",
-                    "yellow",
-                )
-                needsFallback = True
-        else:
-            needsFallback = True
-    else:
-        needsFallback = not supportsCuda
-
-    if needsFallback:
-        from .downloadModels import modelsList
-
-        availableModels = modelsList()
-        methodAttributes = [
-            "interpolate_method",
-            "upscale_method",
-            "segment_method",
-            "depth_method",
-            "restore_method",
-            "dedup_method",
-            "obj_detect_method",
-        ]
-
-        methodToFlag = {
-            "interpolate_method": "interpolate",
-            "upscale_method": "upscale",
-            "segment_method": "segment",
-            "depth_method": "depth",
-            "restore_method": "restore",
-            "dedup_method": "dedup",
-            "obj_detect_method": "obj_detect",
-        }
-
-        for attr in methodAttributes:
-            flagName = methodToFlag.get(attr)
-            if flagName and not getattr(args, flagName):
-                continue
-
-            currentMethod = getattr(args, attr)
-
-            if attr == "restore_method" and isinstance(currentMethod, list):
-                adjusted = []
-                for method in currentMethod:
-                    if any(
-                        backend in method.lower()
-                        for backend in ["-directml", "-ncnn", "-tensorrt", "-mps", "-openvino"]
-                    ):
-                        logging.info(
-                            f"{attr} method {method} already using non-default backend"
-                        )
-                        adjusted.append(method)
-                        continue
-
-                    newMethod = adjustMethod(method, availableModels)
-                    if newMethod != method:
-                        logging.info(
-                            f"Adjusted {attr} method from {method} to {newMethod}"
-                        )
-                    adjusted.append(newMethod)
-                setattr(args, attr, adjusted)
-            else:
-                if any(
-                    backend in currentMethod.lower()
-                    for backend in ["-directml", "-ncnn", "-tensorrt", "-mps", "-openvino"]
-                ):
-                    logging.info(
-                        f"{attr} already using non-default backend: {currentMethod}"
-                    )
-                    continue
-
-                newMethod = adjustMethod(currentMethod, availableModels)
-                if newMethod != currentMethod:
-                    logging.info(f"Adjusted {attr} from {currentMethod} to {newMethod}")
-                    setattr(args, attr, newMethod)
-                else:
-                    logging.info(
-                        f"No adjustment for {attr} ({currentMethod} remains unchanged)"
-                    )
-
-        # Motion blur selects its interpolation backend via --moblur_method. The
-        # bare (rife4.x) and -tensorrt forms both route to CUDA and would crash
-        # on a non-CUDA host (RifeCuda allocates a CUDA stream). adjustMethod()
-        # can't rescue them: rife4.25 has no NCNN/DirectML entry in modelsList,
-        # and NCNN is no longer supported. Force the DirectML backend, which
-        # MotionBlurPipeline supports for every --moblur_method model.
-        if getattr(args, "moblur", False):
-            mob = args.moblur_method
-            if not any(backend in mob for backend in ("-directml", "-openvino")):
-                base = mob.replace("-tensorrt", "")
-                args.moblur_method = f"{base}-directml"
-                logging.info(
-                    f"Adjusted moblur_method from {mob} to {args.moblur_method} "
-                    f"(no CUDA available)"
-                )
-
-
-def processURL(args, outputPath):
-    from urllib.parse import urlparse
-    from src.ytdlp import VideoDownloader
-    from .inputOutputHandler import generateOutputName
-
-    result = urlparse(args.input)
-
-    if result.netloc.lower() in ["www.youtube.com", "youtube.com", "youtu.be"]:
-        logging.info("URL is valid and will be used for processing")
-
-        if args.output is None:
-            outputFolder = os.path.join(outputPath, "output")
-            os.makedirs(os.path.join(outputFolder), exist_ok=True)
-            args.output = os.path.join(
-                outputFolder, generateOutputName(args, args.input)
-            )
-        elif os.path.isdir(args.output):
-            outputFolder = args.output
-            os.makedirs(os.path.join(outputFolder), exist_ok=True)
-        else:
-            outputFolder = os.path.dirname(args.output)
-            os.makedirs(os.path.join(outputFolder), exist_ok=True)
-
-        tempOutput = os.path.join(outputFolder, generateOutputName(args, args.input))
-
-        VideoDownloader(args.input, tempOutput, args.encode_method, args.custom_encoder)
-        logAndPrint(
-            f"Video downloaded successfully to {tempOutput}",
-            "green",
-        )
-
-        if not isAnyOtherProcessingMethodEnabled(args):
-            if tempOutput != args.output:
-                os.rename(tempOutput, args.output)
-                logging.info(f"Renamed output to: {args.output}")
-            sys.exit()
-
-        args.input = str(tempOutput)
-        logging.info(f"New input path: {args.input}")
-    else:
-        logging.error(
-            "URL is invalid or not a YouTube URL, please check the URL and try again"
-        )
-        sys.exit()
+from src.cli.parser import (
+    logAndPrint,
+    _supportsColorStdout,
+    str2bool,
+    DidYouMeanArgumentParser,
+    TASHelpFormatter,
+    capabilityMethods,
+    _listMethods,
+    _buildParser,
+    createParser,
+    _addInterpolationOptions,
+    _addUpscalingOptions,
+    _addDedupOptions,
+    _addVideoProcessingOptions,
+    _addMotionBlurOptions,
+    _addSegmentationOptions,
+    _addSceneDetectionOptions,
+    _addDepthOptions,
+    _addEncodingOptions,
+    _addMiscOptions,
+)
+
+from src.cli.validator import (
+    isAnyOtherProcessingMethodEnabled,
+    argumentsChecker,
+    _autoEnableParentFlags,
+    _handleDepthSettings,
+    _configureProcessingSettings,
+    _validateCustomUpscaleModel,
+    _adjustMethodsBasedOnCuda,
+    processURL,
+    _downloadOfflineModels,
+    _loadJsonConfig,
+)
+
+from src.cli.startup import (
+    _handleDependencies,
+    _promptDownloadRequirementsSelection,
+)
+
+__all__ = [
+    "logAndPrint",
+    "_supportsColorStdout",
+    "str2bool",
+    "DidYouMeanArgumentParser",
+    "TASHelpFormatter",
+    "capabilityMethods",
+    "_listMethods",
+    "_buildParser",
+    "createParser",
+    "_addInterpolationOptions",
+    "_addUpscalingOptions",
+    "_addDedupOptions",
+    "_addVideoProcessingOptions",
+    "_addMotionBlurOptions",
+    "_addSegmentationOptions",
+    "_addSceneDetectionOptions",
+    "_addDepthOptions",
+    "_addEncodingOptions",
+    "_addMiscOptions",
+    "isAnyOtherProcessingMethodEnabled",
+    "argumentsChecker",
+    "_autoEnableParentFlags",
+    "_handleDepthSettings",
+    "_configureProcessingSettings",
+    "_validateCustomUpscaleModel",
+    "_adjustMethodsBasedOnCuda",
+    "processURL",
+    "_downloadOfflineModels",
+    "_loadJsonConfig",
+    "_handleDependencies",
+    "_promptDownloadRequirementsSelection",
+]
